# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FastShip API V1",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/driver/rating": {
            "get": {
                "tags": [
                    "Driver"
                ],
                "summary": "Get all rating driver",
                "operationId": "0189d729d89ba1f04dabc5f51ab501cd",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "user_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "keywords",
                        "in": "query",
                        "description": "keywords",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "star",
                        "in": "query",
                        "description": "star",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all rating"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/driver/cars": {
            "get": {
                "tags": [
                    "Driver"
                ],
                "summary": "Get all cars driver",
                "operationId": "18dac120a3a85c082cfaecac32b59b5f",
                "parameters": [
                    {
                        "name": "keywords",
                        "in": "query",
                        "description": "keywords",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all cars driver"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/driver/payment_method": {
            "get": {
                "tags": [
                    "Driver"
                ],
                "summary": "Get all payment method",
                "operationId": "6db463b599d8c4c257b9b62e047258ee",
                "responses": {
                    "200": {
                        "description": "Get all payment"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/driver/steps": {
            "get": {
                "tags": [
                    "Driver"
                ],
                "summary": "Get all steps",
                "operationId": "203069b7c43ea9575870f5383dd95bc0",
                "responses": {
                    "200": {
                        "description": "Get all steps driver"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/driver/steps/confirm": {
            "post": {
                "tags": [
                    "Driver"
                ],
                "summary": "Confirm step",
                "operationId": "29ae948be2b635b4ce0ef896b0015b20",
                "requestBody": {
                    "description": "Confirm step",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "step_id": {
                                        "description": "ID step confirm",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "image": {
                                        "description": "Image if exits",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Confirm successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/driver/rating/insert": {
            "post": {
                "tags": [
                    "Driver"
                ],
                "summary": "Rating driver",
                "operationId": "ec345ecb19ec9ff7a944f4a70ba9df36",
                "requestBody": {
                    "description": "Rating product with optional images and videos",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "star": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "abcd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rating successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/driver/upload": {
            "post": {
                "tags": [
                    "Driver"
                ],
                "summary": "Upload image",
                "operationId": "dee57768b0b166e27c6c57c0e65a0a0d",
                "requestBody": {
                    "description": "Upload image file",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image",
                                    "type"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "File image upload",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "type": {
                                        "description": "image_cccd_before, image_cccd_after, image_license_before, image_license_after, avatar...",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Upload successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}